# docs/contracts/engineering_openapi.yaml
{
  "components": {
    "schemas": {
...
      "DocumentManifest": {
        "additionalProperties": false,
        "properties": {
          "byte_size": {
            "minimum": 0.0,
            "title": "Byte Size",
            "type": "integer"
          },
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "document_type": {
            "enum": [
              "report_json",
              "report_html",
              "estimate_xlsx"
            ],
            "title": "Document Type",
            "type": "string"
          },
          "download_url": {
            "title": "Download Url",
            "type": "string"
          },
          "file_name": {
            "title": "File Name",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "document_type",
          "file_name",
          "content_type",
          "download_url",
          "byte_size"
        ],
        "title": "DocumentManifest",
        "type": "object"
      },
      "EngineeringPackageRequest": {
        "additionalProperties": false,
        "properties": {
          "documents": {
            "items": {
              "enum": [
                "report_json",
                "report_html",
                "estimate_xlsx"
              ],
              "type": "string"
            },
            "title": "Documents",
            "type": "array"
          },
          "revision": {
            "pattern": "^[A-Za-z0-9._-]{1,80}$",
            "title": "Revision",
            "type": "string"
          }
        },
        "required": [
          "revision"
        ],
        "title": "EngineeringPackageRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobStatus": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentManifest"
            },
            "title": "Documents",
            "type": "array"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "package_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Package Id"
          },
          "progress": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress",
            "type": "integer"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "string"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          },
          "status": {
            "enum": [
              "queued",
              "processing",
              "completed",
              "completed_with_warnings",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "project_id",
          "revision",
          "status",
          "progress",
          "stage"
        ],
        "title": "JobStatus",
        "type": "object"
      },
      "LockRevisionRequest": {
        "additionalProperties": false,
        "properties": {
          "confirmed_by": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Confirmed By",
            "type": "string"
          }
        },
        "required": [
          "confirmed_by"
        ],
        "title": "LockRevisionRequest",
        "type": "object"
      },
...
      "ReportPayload": {
        "additionalProperties": false,
        "properties": {
          "assumptions": {
            "items": {
              "type": "string"
            },
            "title": "Assumptions",
            "type": "array"
          },
          "demo_disclaimer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Demo Disclaimer"
          },
          "demo_mode": {
            "title": "Demo Mode",
            "type": "boolean"
          },
          "design_narrative": {
            "$ref": "#/components/schemas/DesignNarrative"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentManifest"
            },
            "title": "Documents",
            "type": "array"
          },
          "estimate": {
            "$ref": "#/components/schemas/EstimateResult"
          },
          "exclusions": {
            "items": {
              "type": "string"
            },
            "title": "Exclusions",
            "type": "array"
          },
          "furniture_estimate": {
            "$ref": "#/components/schemas/FurnitureEstimateResult"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "narratives": {
            "$ref": "#/components/schemas/NarrativeResult"
          },
          "package_id": {
            "title": "Package Id",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "quantities": {
            "$ref": "#/components/schemas/QuantityResult"
          },
          "retrieval": {
            "$ref": "#/components/schemas/RetrievalResult"
          },
          "revision": {
            "title": "Revision",
            "type": "string"
          },
          "risks": {
            "$ref": "#/components/schemas/RiskResult"
          },
          "schedule": {
            "$ref": "#/components/schemas/ScheduleResult"
          },
          "schema_version": {
            "const": "roompilot.report-payload.v1",
            "default": "roompilot.report-payload.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "snapshot": {
            "$ref": "#/components/schemas/ProjectSnapshot"
          },
          "snapshot_hash": {
            "title": "Snapshot Hash",
            "type": "string"
          },
          "status": {
            "enum": [
              "completed",
              "completed_with_warnings"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "package_id",
          "project_id",
          "revision",
          "generated_at",
          "status",
          "snapshot_hash",
          "demo_mode",
          "demo_disclaimer",
          "snapshot",
          "quantities",
          "retrieval",
          "risks",
          "estimate",
          "furniture_estimate",
          "schedule",
          "narratives",
          "design_narrative",
          "assumptions",
          "exclusions"
        ],
        "title": "ReportPayload",
        "type": "object"
      },
...
      "SnapshotEnvelope": {
        "additionalProperties": false,
        "properties": {
          "completeness": {
            "$ref": "#/components/schemas/SnapshotCompleteness"
          },
          "snapshot": {
            "$ref": "#/components/schemas/ProjectSnapshot"
          }
        },
        "required": [
          "snapshot",
          "completeness"
        ],
        "title": "SnapshotEnvelope",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "由目前 FastAPI route 與 Pydantic 模型輸出的工程文件 API；前端不得直連 LLM、Vector Index 或資料庫。",
    "title": "RoomPilot Designer-Locked Engineering API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/documents/{document_id}/download": {
      "get": {
        "operationId": "download_document_api_v1_documents__document_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "preview",
            "required": false,
            "schema": {
              "default": false,
              "title": "Preview",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Document",
        "tags": [
          "engineering-documents"
        ]
      }
    },
...
    "/api/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "engineering-documents"
        ]
      }
    },
    "/api/v1/packages/{package_id}": {
      "get": {
        "operationId": "get_package_api_v1_packages__package_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "package_id",
            "required": true,
            "schema": {
              "title": "Package Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportPayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Package",
        "tags": [
          "engineering-documents"
        ]
      }
    },
    "/api/v1/projects/{project_id}/engineering-packages": {
      "post": {
        "operationId": "create_engineering_package_api_v1_projects__project_id__engineering_packages_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EngineeringPackageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Engineering Package",
        "tags": [
          "engineering-documents"
        ]
      }
    },
    "/api/v1/projects/{project_id}/revisions/{revision}/lock": {
      "post": {
        "operationId": "lock_revision_api_v1_projects__project_id__revisions__revision__lock_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision",
            "required": true,
            "schema": {
              "title": "Revision",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LockRevisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lock Revision",
        "tags": [
          "engineering-documents"
        ]
      }
    },
    "/api/v1/projects/{project_id}/revisions/{revision}/snapshot": {
      "get": {
        "operationId": "get_snapshot_api_v1_projects__project_id__revisions__revision__snapshot_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision",
            "required": true,
            "schema": {
              "title": "Revision",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Snapshot",
        "tags": [
          "engineering-documents"
        ]
      },
      "put": {
        "operationId": "save_snapshot_api_v1_projects__project_id__revisions__revision__snapshot_put",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision",
            "required": true,
            "schema": {
              "title": "Revision",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectSnapshot"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Snapshot",
        "tags": [
          "engineering-documents"
        ]
      }
    }
  }
}